requirements.txt
anthropic
requests
pandas
python-dotenv
beautifulsoup4
lxml
tqdm
jupyter
